Bsd 2181 better handling of tax id of samples

+++ pipelines/ebeye-search-export/src/main/java/uk/ac/ebi/biosamples/ebeye/base/EbEyeBioSamplesDataDumpRunner.java
@@ -272,7 +272,7 @@ public class EbEyeBioSamplesDataDumpRunner implements ApplicationRunner {
     return crossReferencesType;
   }
 
-  private RefType getTaxonomyCrossReferenceCovid(int taxId) {
+  private RefType getTaxonomyCrossReferenceCovid(long taxId) {
     RefType refType = new RefType();
 
     refType.setDbname("TAXONOMY");
